# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Uplift Core API
  description: |
    Workforce management platform API for scheduling, time tracking, and employee management.
    
    ## Authentication
    All endpoints except `/auth/login`, `/auth/register`, and `/health` require authentication.
    Authentication uses httpOnly cookies set by the login endpoint.
    
    ## CSRF Protection
    Non-GET requests require a CSRF token. Get a token from `GET /csrf-token` and include it in the `X-CSRF-Token` header.
    
    ## Rate Limiting
    - Auth endpoints: 10 requests per 15 minutes
    - API endpoints: 100 requests per 15 minutes
    
  version: 1.0.0
  contact:
    name: Uplift Support
    email: support@uplift.hr
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://api.uplifthq.co.uk/api
    description: Production
...
paths:
...
  /employees/{id}:
    get:
      summary: Get employee
      tags: [Employees]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee:
                    $ref: '#/components/schemas/Employee'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update employee
      tags: [Employees]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '200':
          description: Employee updated
    delete:
      summary: Delete employee
      tags: [Employees]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Employee deleted
...
  /dashboard:
    get:
      summary: Get dashboard data
      tags: [Dashboard]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Dashboard metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  today:
                    type: object
                  week:
                    type: object
                  pending:
                    type: object
...
  /momentum/me:
    get:
      summary: Get my momentum score
      tags: [Momentum]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user's momentum score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MomentumScore'
  /momentum/dashboard:
    get:
      summary: Get organization momentum dashboard
      tags: [Momentum]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Momentum dashboard metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  orgAverage:
                    type: number
                  distribution:
                    type: array
                  trends:
                    type: object
  /momentum/leaderboard:
    get:
      summary: Get momentum leaderboard
      tags: [Momentum]
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Top employees by momentum score
          content:
            application/json:
              schema:
                type: object
                properties:
                  leaderboard:
                    type: array
                    items:
                      $ref: '#/components/schemas/MomentumScore'

  # ==================== PAYROLL ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
...
    Employee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        employeeNumber:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, casual, contractor]
        status:
          type: string
          enum: [active, inactive, terminated]
        startDate:
          type: string
          format: date
        departmentId:
          type: string
          format: uuid
        primaryLocationId:
          type: string
          format: uuid
        hourlyRate:
          type: number
    EmployeeInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        employmentType:
          type: string
        startDate:
          type: string
          format: date
        departmentId:
          type: string
          format: uuid
        primaryLocationId:
          type: string
          format: uuid
        hourlyRate:
          type: number
...
    MomentumScore:
      type: object
      properties:
        employeeId:
          type: string
          format: uuid
        employeeName:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        trend:
          type: string
          enum: [up, down, stable]
        breakdown:
          type: object
          properties:
            attendance:
              type: number
            punctuality:
              type: number
            shiftCompletion:
              type: number
            skillsGrowth:
              type: number
            recognition:
              type: number
            engagement:
              type: number
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
...